requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
pymongo==4.6.0
pandas==2.1.4
numpy==1.24.3
lxml==4.9.3
python-dotenv==1.0.0
fake-useragent==1.4.0
time
re
json
urllib.parse